[docs] SCRD-6795 Create ardana-dev-tools walkthrough
This walkthrough document was created from the experiences of trying to
create and work with the ardana-dev-tools project. Many of these steps
can be found in the ardana-dev-tools README.md, but many more were
obtained though emails and chat conversations with various experts. In
order to help future developers avoid the pitfalls experienced, this
document was created.

Change-Id: I1fca85c77a33613e8202ce4728b40bbc4a502041
Signed-off-by: John L. Jolly <jjolly@suse.com>